# openapi.yaml
openapi: 3.1.0
info:
  title: Chat With Docs API
  version: 0.1.0
  description: Local-first document ingestion and page-aware RAG API.
servers:
  - url: http://localhost:3000
paths:
...
  /v1/documents:
    get:
      summary: List documents
      responses:
        '200': { description: Documents }
    post:
      summary: Upload a PDF or DOCX and enqueue processing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, mimeType, contentBase64]
              properties:
                filename: { type: string }
                mimeType: { type: string, enum: [application/pdf, application/vnd.openxmlformats-officedocument.wordprocessingml.document] }
                contentBase64: { type: string, format: byte }
      responses:
        '202': { description: Processing job accepted }
...
  /v1/documents/{documentId}:
    delete:
      summary: Delete a document and its indexed content
      parameters:
        - { name: documentId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Document deleted }
        '404': { description: Document not found }
...